% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_RGtk2_GUI.R
\name{askQuit}
\alias{askQuit}
\title{RGtk2 GUI function: ask confirmation to quit if topright button is used to quit.}
\usage{
askQuit(myobject)
}
\arguments{
\item{myobject}{The parent window to be closed.}
}
\description{
This function is provided so that you can easily see its content. It is not
  intended to be used, prefer \code{loadGUI()} to load the \code{RGtk2} GUI.
}

