% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_RGtk2_GUI.R
\name{checkEntry}
\alias{checkEntry}
\title{RGtk2 GUI function: check data validity in entries.}
\usage{
checkEntry(validatedEntry, myRegEx)
}
\arguments{
\item{validatedEntry}{Entry to be checked.}

\item{myRegEx}{Regular expression to test the \code{validatedEntry}.}
}
\description{
This function is provided so that you can easily see its content. It is not
  intended to be used, prefer \code{loadGUI()} to load the \code{RGtk2} GUI.
}

