% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_ANA_correspAnalysis.R
\name{doCA}
\alias{doCA}
\title{Performs a correspondance analysis on the basis of the word-occurrence data.frame.}
\usage{
doCA(wordF, getPlot = TRUE, mwidth = 800, mheight = 800, ...)
}
\arguments{
\item{wordF}{The data.frame containing word occurrences.}

\item{getPlot}{If \code{TRUE}, save the \code{\link[ca]{ca}} plot in the RESULTS
directory.}

\item{mwidth}{The width of the plot in pixels.}

\item{mheight}{The height of the plot in pixels.}

\item{...}{Additional arguments from the \code{\link[ca]{ca}} function.}
}
\value{
The results of the \code{\link[ca]{ca}} function.
}
\description{
Performs a correspondance analysis on the basis of the word-occurrence data.frame
   using \code{\link[ca]{ca}} function.
}
\examples{
\dontrun{
doCA(wordF=myDF)
}
}

