% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_dataset_docu.R
\docType{data}
\name{exclusionList_FR}
\alias{exclusionList_FR}
\title{Stop words in French.}
\format{A vector with 173 elements (character), with UTF-8 characters escaped
  using \code{stringi::stri_escape_unicode(exclusionList_FR)}.}
\source{
Adapted from www.ranks.nl/stopwords.
}
\usage{
exclusionList_FR
}
\description{
A vector containing stop words in French.
}
\keyword{datasets}
