% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_dataset_docu.R
\docType{data}
\name{exclusionList_SP}
\alias{exclusionList_SP}
\title{Stop words in Spanish.}
\format{A vector with 190 elements (character), with UTF-8 characters escaped
  using \code{stringi::stri_escape_unicode(exclusionList_SP)}.}
\source{
Adapted from \url{http://www.ranks.nl/stopwords/spanish}.
}
\usage{
exclusionList_SP
}
\description{
A vector containing stop words in Spanish
}
\keyword{datasets}

