% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_dataset_docu.R
\docType{data}
\name{exclusionList_UK}
\alias{exclusionList_UK}
\title{Stop words in English.}
\format{
A vector with 542 elements (character).
}
\source{
Adapted from www.ranks.nl/stopwords.
}
\usage{
exclusionList_UK
}
\description{
A vector containing stop words in English.
}
\keyword{datasets}
