% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_ALL_quickFunctions.R
\name{getAllAnalysis}
\alias{getAllAnalysis}
\title{A quick way to compute a set of analysis from the word-occurrence data.frame.}
\usage{
getAllAnalysis(dataset)
}
\arguments{
\item{dataset}{A single word-occurrrence data.frame.}
}
\value{
A set of analyses available from the \code{inpdfr} package.
}
\description{
A quick way to compute a set of analysis from the word-occurrence data.frame.
}
\examples{
\dontrun{
getAllAnalysis(dataset=myDF)
}
}

