% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_PRO_manageFiles.R
\name{getListFiles}
\alias{getListFiles}
\title{List files in a specified directory sorted by extension.}
\usage{
getListFiles(mywd)
}
\arguments{
\item{mywd}{A string containing the working directory.}
}
\value{
A list of length 2 with file names sorted by extension (pdf and txt).
}
\description{
List files in a specified directory sorted by extension. The function takes
  into account .txt and .pdf files based on \code{strsplit} function.
}
\examples{
getListFiles(mywd = getwd())
}

