% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_PRO_stopwords.R
\name{getStopWords}
\alias{getStopWords}
\title{Load a list of stopwords.}
\usage{
getStopWords()
}
\value{
A list of vectors with stopwords for French, English, and Spanish languages.
}
\description{
\code{getStopWords} returns a list of stopwords.
}
\examples{
getStopWords()
}
