% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_ANA_occur.R
\name{getSummaryStatsOCCUR}
\alias{getSummaryStatsOCCUR}
\title{Plot a scatter plot with the proportion of documents using similar words.}
\usage{
getSummaryStatsOCCUR(
  wordF,
  getPlot = TRUE,
  mwidth = 800,
  mheight = 800,
  formatType = "png"
)
}
\arguments{
\item{wordF}{The data.frame containing word occurrences.}

\item{getPlot}{If \code{TRUE}, save the scatter plot in the RESULTS directory.}

\item{mwidth}{The width of the plot in pixels.}

\item{mheight}{The height of the plot in pixels.}

\item{formatType}{The format for the output file ("eps", "pdf", "png", "svg", "tiff", "jpeg", "bmp").}
}
\value{
A data.frame containing the proportion of documents and the number of similar words.
}
\description{
Plot a scatter plot with the proportion of documents using similar words.
}
\examples{
\dontrun{
getSummaryStatsOCCUR(wordF = myDF)
}
}
