% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_PRO_extractTxt.R
\name{getTXT}
\alias{getTXT}
\title{Extract text from TXT files and return a word-occurrence data.frame.}
\usage{
getTXT(myTXTs)
}
\arguments{
\item{myTXTs}{A character vector containing TXT file names (or complete path
to these files).}
}
\value{
A list of list with word-occurrence data.frame and file name.
}
\description{
Extract text from TXT files and return a word-occurrence data.frame.
}
\examples{
\dontrun{
getTXT(myTXTs="mytxt.txt")
}
}

