% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_ALL_quickFunctions.R
\name{getwordOccuDF}
\alias{getwordOccuDF}
\title{A quick way to obtain the word-occurrence data.frame from a set of documents.}
\usage{
getwordOccuDF(mywd, language = "English")
}
\arguments{
\item{mywd}{A character variable containing the working directory.}

\item{language}{The language used ("French", "English", "Spanish").}
}
\value{
A single word-occurrrence data.frame.
}
\description{
A quick way to obtain the word-occurrence data.frame from a set of documents.
}
\examples{
\dontrun{
getwordOccuDF(mywd=getwd(),language="English")
}
}

