% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_ANA_freqWordCor.R
\name{getXFreqWord}
\alias{getXFreqWord}
\title{Returns most frequent words}
\usage{
getXFreqWord(wordF, occuWords)
}
\arguments{
\item{wordF}{The data.frame containing word occurrences.}

\item{occuWords}{The minimum number of occurrences for words to be returned.}
}
\value{
A vector with most frequent words.
}
\description{
Returns most frequent words
}
\examples{
\dontrun{
getXFreqWord(wordF=myDF,occuWords=5)
}
}

