% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_ANA_ECO_metacom_identifyStructure.R
\name{IdentifyStructure}
\alias{IdentifyStructure}
\title{Copy of the identifyStructure function from Tad Dallas metacom package.}
\usage{
IdentifyStructure(metacom.obj)
}
\arguments{
\item{metacom.obj}{The result of the `Metacommunity` function, containing a 
list of 4 elements; the empirical matrix being tested, and results for 
coherence, turnover, and boundary clumping.}
}
\value{
Ouputs a classification of the metacommunity.
}
\description{
Identifies structure (or quasi-structure) and outputs a classification.
}
\details{
Tad Dallas <tdallas@uga.edu> identifyStructure function no longer 
  maintained in metacom package. see https://github.com/taddallas/metacom. 
  This function was copy-pasted from version 1.4.4 of package metacom with 
  minor modification (fix warning: the condition has length > 1 and only the 
  first element will be used).
}
\note{
Quasi structures, as well as 'random' and 'Gleasonian' structures, 
  may not strictly be discernable through the EMS approach, as they rely on 
  inferring a result from a non-significant test ('accepting the null'), 
  which is typically a bad idea.
}
