% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_dataset_docu.R
\docType{data}
\name{loremIpsum}
\alias{loremIpsum}
\title{Lorem Ipsum text.}
\format{A vector with 556 elements, each element corresponds to a line in the original text (character).}
\source{
\url{http://lipsum.com/}.
}
\usage{
loremIpsum
}
\description{
A vector containing a Lorem Ipsum text for testing purposes.
}
\keyword{datasets}

