% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_RGtk2_GUI.R
\name{makeMenuMainWindow}
\alias{makeMenuMainWindow}
\title{RGtk2 GUI function: main window.}
\usage{
makeMenuMainWindow(main_window)
}
\arguments{
\item{main_window}{Main window where the menu is created.}
}
\description{
This function is provided so that you can easily see its content. It is not
  intended to be used, prefer \code{loadGUI()} to load the \code{RGtk2} GUI.
}

