% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_RGtk2_GUI.R
\name{open_cb}
\alias{open_cb}
\title{RGtk2 GUI function: open a window in order to choose the working directory.}
\usage{
open_cb(widget, window)
}
\arguments{
\item{widget}{Widget to open.}

\item{window}{A window to contain the widget.}
}
\value{
The path to the user-defeined working directory.
}
\description{
This function is provided so that you can easily see its content. It is not
  intended to be used, prefer \code{loadGUI()} to load the \code{RGtk2} GUI.
}

