% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_PRO_manageFiles.R
\name{quitSpaceFromChars}
\alias{quitSpaceFromChars}
\title{Delete spaces in file names.}
\usage{
quitSpaceFromChars(vectxt)
}
\arguments{
\item{vectxt}{A vector containing character entries corresponding to the names
of files in the current working directory.}
}
\value{
The function returns a logical for each file, with TRUE if the file
  has been found, and FALSE otherwise.
}
\description{
Delete spaces in file names located in the current working directory.
}
\examples{
\dontrun{
quitSpaceFromChars(c("my pdf.pdf","my other pdf.pdf"))
}
}
