% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_RGtk2_GUI.R
\name{switchOffDialogWait}
\alias{switchOffDialogWait}
\title{RGtk2 GUI function: switch off the "Processing..." message.}
\usage{
switchOffDialogWait(dialogX)
}
\arguments{
\item{dialogX}{The dialog window to be closed.}
}
\description{
This function is provided so that you can easily see its content. It is not
  intended to be used, prefer \code{loadGUI()} to load the \code{RGtk2} GUI.
}

