% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_RGtk2_GUI.R
\name{switchOnDialogWait}
\alias{switchOnDialogWait}
\title{RGtk2 GUI function: switch on the "Processing..." message.}
\usage{
switchOnDialogWait()
}
\description{
This function is provided so that you can easily see its content. It is not
  intended to be used, prefer \code{loadGUI()} to load the \code{RGtk2} GUI.
}

