% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_PRO_processDF.R
\name{truncNumWords}
\alias{truncNumWords}
\title{Truncate the word-occurrence data.frame.}
\usage{
truncNumWords(wordF, maxWords)
}
\arguments{
\item{wordF}{The data.frame containing word occurrences.}

\item{maxWords}{The maximum number of words in the data.frame.}
}
\value{
The data.frame containing word occurrences.
}
\description{
Truncate the word-occurrence data.frame.
}
\examples{
\dontrun{
truncNumWords(wordF = myWordOccurrenceDF, maxWords = 50)
}
}
