% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_dataset_docu.R
\docType{data}
\name{wordOccuDF}
\alias{wordOccuDF}
\title{Lorem Ipsum word occurrences.}
\format{
A data.frame containing word name and occurences for testing purposes.
}
\usage{
wordOccuDF
}
\description{
Lorem Ipsum word occurrences.
}
\keyword{datasets}
