% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_ANA_freqWordCor.R
\name{getMostFreqWord}
\alias{getMostFreqWord}
\title{Returns most frequent words.}
\usage{
getMostFreqWord(
  wordF,
  numWords,
  getPlot = TRUE,
  mwidth = 1024,
  mheight = 800,
  formatType = "png"
)
}
\arguments{
\item{wordF}{The data.frame containing word occurrences.}

\item{numWords}{The number of words to be returned.}

\item{getPlot}{If \code{TRUE}, save a scatter plot in the RESULTS directory.}

\item{mwidth}{The width of the plot in pixels.}

\item{mheight}{The height of the plot in pixels.}

\item{formatType}{The format for the output file ("eps", "pdf", "png", "svg", "tiff", "jpeg", "bmp").}
}
\value{
The \code{numWords} most frequent words.
}
\description{
Returns most frequent words and plots their frequencies per document.
}
\examples{
data("wordOccuDF")
getMostFreqWord(wordF = wordOccuDF, numWords = 5, getPlot = FALSE)
}
