% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_ANA_barplot.R
\name{getSummaryStatsBARPLOT}
\alias{getSummaryStatsBARPLOT}
\title{Perform a barplot with the number of unique words per document}
\usage{
getSummaryStatsBARPLOT(wordF, getPlot = TRUE, mwidth = 800, mheight = 800,
  ...)
}
\arguments{
\item{wordF}{The data.frame containing word occurrences.}

\item{getPlot}{If \code{TRUE}, save the bar plot in the RESULTS directory.}

\item{mwidth}{The width of the plot in pixels.}

\item{mheight}{The height of the plot in pixels.}

\item{...}{Additional arguments from \code{barplot} function.}
}
\value{
The number of unique words per document.
}
\description{
Perform a barplot with the number of unique words per document using \code{\link[graphics]{barplot}} function.
}
\examples{
\dontrun{
getSummaryStatsBARPLOT(wordF=myDF)
}
}

