% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpdfr_PRO_processDF.R
\name{mergeWordFreq}
\alias{mergeWordFreq}
\title{Merge word-occurrence data.frames into a single data.frame.}
\usage{
mergeWordFreq(wordF)
}
\arguments{
\item{wordF}{The data.frame containing word occurrences.}
}
\value{
A single word-occurrrence data.frame with each column corresponding to a text file.
}
\description{
Merge word-occurrence data.frames into a single data.frame.
}
\examples{
\dontrun{
mergeWordFreq(wordF=myWordOccurrenceDF)
}
}

